A couple more minor signedness fixes

+++ src/moar.c
@@ -481,7 +481,7 @@ void MVM_vm_dump_file(MVMInstance *instance, const char *filename) {
     MVMCompUnit      *cu = MVM_cu_map_from_file(tc, filename);
     char *dump = MVM_bytecode_dump(tc, cu);
     size_t dumplen = strlen(dump);
-    int position = 0;
+    size_t position = 0;
 
     /* libuv already set up stdout to be nonblocking, but it can very well be
      * we encounter EAGAIN (Resource temporarily unavailable), so we need to
